1. optimize adcs7476 data process 2. modify plan data time parameter 3. optimize trigger out signal

+++ ArmCoreV1_QSPI/application/app/main_app/plan_data.h
@@ -13,7 +13,7 @@ struct radiation_point_data
 {
     float dose_cumulative;
     float dose_rate;        /* Unit: MU/min */
-    uint16_t time_expected; /* Unit: ms */
+    uint32_t time_expected; /* Unit: ms */
 };
 
 #define BEAMS_MAX           1
